Add initial updates for Python3.8.10 compatibility
Most of the updates are quite minor, in retrospect. The handling of what
are now raw bytes (previously strings) proved to be the most
challenging. Of course, other than attempting to and failing at getting
upload_to_ram_and_program working for SKARAB.

A lot of debug has been left in transport_skarab.py and
skarab_fileops.py for now, I am currently using a workaround of manually
programming boards via the progska utility and running through the
reboot steps.

This branch's development is largely to facilitate DSim usage and
testing for katgpucbf.

Contributes to: NGC-393.